starting dexie work
