--- app/FilesApp/Layout.tsx
@@ -1,141 +1,0 @@
-import { Box, BoxProps, Paper } from '@mui/material'
-
-const Root = (props: BoxProps) => (
-  <Box
-    {...props}
-    sx={[
-      {
-        bgcolor: 'background.surface',
-        display: 'grid',
-        gridTemplateColumns: {
-          xs: '1fr',
-          sm: 'minmax(64px, 200px) minmax(450px, 1fr)',
-          md: 'minmax(160px, 250px) minmax(600px, 1fr)',
-        },
-        gridTemplateRows: '64px 1fr',
-        minHeight: '100vh',
-      },
-      ...(Array.isArray(props.sx) ? props.sx : [props.sx]),
-    ]}
-  />
-)
-
-const Header = (props: BoxProps) => (
-  <Box
-    component="header"
-    className="Header"
-    {...props}
-    sx={[
-      {
-        p: 2,
-        gap: 2,
-        bgcolor: 'background.surface',
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        gridColumn: '1 / -1',
-        borderBottom: '1px solid',
-        borderColor: 'divider',
-        position: 'sticky',
-        top: 0,
-        zIndex: 1100,
-      },
-      ...(Array.isArray(props.sx) ? props.sx : [props.sx]),
-    ]}
-  />
-)
-
-const SideNav = (props: BoxProps) => (
-  <Box
-    component="nav"
-    className="Navigation"
-    {...props}
-    sx={[
-      {
-        p: 2,
-        bgcolor: 'background.surface',
-        borderRight: '1px solid',
-        borderColor: 'divider',
-        display: {
-          xs: 'none',
-          sm: 'initial',
-        },
-      },
-      ...(Array.isArray(props.sx) ? props.sx : [props.sx]),
-    ]}
-  />
-)
-
-const SidePane = (props: BoxProps) => (
-  <Box
-    className="Inbox"
-    {...props}
-    sx={[
-      {
-        bgcolor: 'background.surface',
-        borderRight: '1px solid',
-        borderColor: 'divider',
-        display: {
-          xs: 'none',
-          md: 'initial',
-        },
-      },
-      ...(Array.isArray(props.sx) ? props.sx : [props.sx]),
-    ]}
-  />
-)
-
-const Main = (props: BoxProps) => (
-  <Box
-    component="main"
-    className="Main"
-    {...props}
-    sx={[{ p: 2 }, ...(Array.isArray(props.sx) ? props.sx : [props.sx])]}
-  />
-)
-
-const SideDrawer = ({
-  onClose,
-  ...props
-}: BoxProps & { onClose: React.MouseEventHandler<HTMLDivElement> }) => (
-  <Box
-    {...props}
-    sx={[
-      { position: 'fixed', zIndex: 1200, width: '100%', height: '100%' },
-      ...(Array.isArray(props.sx) ? props.sx : [props.sx]),
-    ]}
-  >
-    <Box
-      role="button"
-      onClick={onClose}
-      sx={{
-        position: 'absolute',
-        inset: 0,
-        bgcolor: theme =>
-          `rgba(${theme.vars.palette.neutral.darkChannel} / 0.8)`,
-      }}
-    />
-    <Paper
-      sx={{
-        minWidth: 256,
-        width: 'max-content',
-        height: '100%',
-        p: 2,
-        boxShadow: 'lg',
-        bgcolor: 'background.surface',
-      }}
-    >
-      {props.children}
-    </Paper>
-  </Box>
-)
-
-export default {
-  Root,
-  Header,
-  SideNav,
-  SidePane,
-  SideDrawer,
-  Main,
-}
