+++ app/components/DarkMode.tsx
@@ -1,19 +1,9 @@
 import { useColorScheme, IconButton } from '@mui/joy'
-import { useState, useEffect } from 'react'
 import { DarkModeRounded, LightModeRounded } from '@mui/icons-material'
 
-//const light = new URL('../assets/light.mp3', import.meta.url)
-
-export const DarkMode: React.FunctionComponent = () => {
+export default function DarkMode() {
   const { mode, setMode } = useColorScheme()
-  const [mounted, setMounted] = useState(false)
 
-  useEffect(() => {
-    setMounted(true)
-  }, [])
-  if (!mounted) {
-    return <IconButton size="sm" variant="outlined" color="primary" />
-  }
   return (
     <IconButton
       id="toggle-mode"
@@ -21,7 +11,7 @@ export const DarkMode: React.FunctionComponent = () => {
       variant="outlined"
       color="primary"
       onClick={() => {
-        //new Audio(light.href).play()
+        new Audio('/light.mp3').play()
         if (mode === 'light') {
           setMode('dark')
         } else {
