+++ app/components/layout/LeftNav.tsx
@@ -1,0 +1,207 @@
+import {
+  Box,
+  Typography,
+  IconButton,
+  List,
+  ListItem,
+  ListItemButton,
+  ListItemDecorator,
+  ListItemContent,
+} from '@mui/joy'
+
+// Icons import
+import {
+  FolderOpen,
+  ShareOutlined,
+  KeyboardArrowDownRounded,
+  DeleteRounded,
+} from '@mui/icons-material'
+
+export default function LeftNav() {
+  return (
+    <Box
+      component="nav"
+      className="Navigation"
+      sx={[
+        {
+          p: 2,
+          bgcolor: 'background.surface',
+          borderRight: '1px solid',
+          borderColor: 'divider',
+          display: {
+            xs: 'none',
+            sm: 'initial',
+          },
+        },
+      ]}
+    >
+      <List size="sm" sx={{ '--List-item-radius': '8px' }}>
+        <ListItem nested sx={{ p: 0 }}>
+          <Box
+            sx={{
+              mb: 1,
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'space-between',
+            }}
+          >
+            <Typography
+              id="nav-list-browse"
+              textColor="neutral.500"
+              fontWeight={700}
+              sx={{
+                fontSize: '10px',
+                textTransform: 'uppercase',
+                letterSpacing: '.1rem',
+              }}
+            >
+              Browse
+            </Typography>
+            <IconButton
+              size="sm"
+              variant="plain"
+              color="primary"
+              sx={{ '--IconButton-size': '24px' }}
+            >
+              <KeyboardArrowDownRounded fontSize="small" color="primary" />
+            </IconButton>
+          </Box>
+          <List
+            aria-labelledby="nav-list-browse"
+            sx={{
+              '& .JoyListItemButton-root': { p: '8px' },
+            }}
+          >
+            <ListItem>
+              <ListItemButton variant="soft" color="primary">
+                <ListItemDecorator sx={{ color: 'inherit' }}>
+                  <FolderOpen fontSize="small" />
+                </ListItemDecorator>
+                <ListItemContent>My files</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator sx={{ color: 'neutral.500' }}>
+                  <ShareOutlined fontSize="small" />
+                </ListItemDecorator>
+                <ListItemContent>
+                  <ListItemContent>Shared files</ListItemContent>
+                </ListItemContent>
+              </ListItemButton>
+            </ListItem>
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator sx={{ color: 'neutral.500' }}>
+                  <DeleteRounded fontSize="small" />
+                </ListItemDecorator>
+                <ListItemContent>Trash</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+          </List>
+        </ListItem>
+        <ListItem nested>
+          <Box
+            sx={{
+              mt: 2,
+              mb: 1,
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'space-between',
+            }}
+          >
+            <Typography
+              id="nav-list-tags"
+              textColor="neutral.500"
+              fontWeight={700}
+              sx={{
+                fontSize: '10px',
+                textTransform: 'uppercase',
+                letterSpacing: '.1rem',
+              }}
+            >
+              Tags
+            </Typography>
+            <IconButton
+              size="sm"
+              variant="plain"
+              color="primary"
+              sx={{ '--IconButton-size': '24px' }}
+            >
+              <KeyboardArrowDownRounded fontSize="small" color="primary" />
+            </IconButton>
+          </Box>
+          <List
+            aria-labelledby="nav-list-tags"
+            size="sm"
+            sx={{
+              '--List-decorator-size': '32px',
+              '& .JoyListItemButton-root': { p: '8px' },
+            }}
+          >
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator>
+                  <Box
+                    sx={{
+                      width: '10px',
+                      height: '10px',
+                      borderRadius: '99px',
+                      bgcolor: 'primary.300',
+                    }}
+                  />
+                </ListItemDecorator>
+                <ListItemContent>Personal</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator>
+                  <Box
+                    sx={{
+                      width: '10px',
+                      height: '10px',
+                      borderRadius: '99px',
+                      bgcolor: 'danger.400',
+                    }}
+                  />
+                </ListItemDecorator>
+                <ListItemContent>Work</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator>
+                  <Box
+                    sx={{
+                      width: '10px',
+                      height: '10px',
+                      borderRadius: '99px',
+                      bgcolor: 'warning.500',
+                    }}
+                  />
+                </ListItemDecorator>
+                <ListItemContent>Travels</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+            <ListItem>
+              <ListItemButton>
+                <ListItemDecorator>
+                  <Box
+                    sx={{
+                      width: '10px',
+                      height: '10px',
+                      borderRadius: '99px',
+                      bgcolor: 'success.400',
+                    }}
+                  />
+                </ListItemDecorator>
+                <ListItemContent>Concert tickets</ListItemContent>
+              </ListItemButton>
+            </ListItem>
+          </List>
+        </ListItem>
+      </List>
+    </Box>
+  )
+}
