+++ app/components/mixes/mixes.tsx
@@ -3,8 +3,8 @@ import { ButtonGroup } from '@mui/material'
 import { Button, Breadcrumbs, Card } from '@mui/joy'
 import { Stop, Pause, PlayArrow, Shuffle } from '@mui/icons-material'
 import TrackForm from './trackform'
-import { Events } from '../../Events'
-import { db, TrackState, useLiveQuery } from '../../db'
+import { Events } from '../../../app/api/Events'
+import { db, TrackState, useLiveQuery } from '../../../app/api/db'
 export const Mixes = () => {
   const [track0, track1]: TrackState[] =
     useLiveQuery(() => db.trackState.limit(2).toArray()) || []
