+++ package.json
@@ -31,8 +31,8 @@
     "peaks.js": "^2.0.6",
     "prop-types": "^15.8.1",
     "rc-slider": "^10.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "waveform-data": "^4.3.0",
     "web-audio-beat-detector": "^8.1.33"
   },
