+++ remix.config.js
@@ -1,1 +1,6 @@
-module.exports = {}
+/** @type {import('@remix-run/dev').AppConfig} */
+module.exports = {
+  appDirectory: 'app', // default
+  ignoredRouteFiles: ['**/.*'], // default
+  assetsBuildDirectory: 'public/build', // default
+}
